drawing circles on map, xyz for 'scanprints' (fingerprints), added kryo for serialisation

+++ app/src/main/java/sorja/indoorlocation/ScanPrint.java
@@ -4,11 +4,11 @@ package sorja.indoorlocation;
  * Created by mxO on 6.12.2015.
  */
 public class ScanPrint extends FingerPrint {
-    int x;
-    int y;
-    int z;
+    float x;
+    float y;
+    float z;
 
-    public ScanPrint(String ssid, int rssi, String mac, long timestamp, int x, int y, int z) {
+    public ScanPrint(String ssid, int rssi, String mac, long timestamp, float x, float y, float z) {
         super(ssid, rssi, mac, timestamp);
         this.x = x;
         this.y = y;
